backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
docx2txt
PyPDF2

